requirements.txt

# -*- coding: utf-8 -*-
pywin32>=227
Pillow>=8.0.0
pynput>=1.7.0
numpy>=1.19.0